Fixed types conversions

+++ srcs/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: astadnik <astadnik@student.unit.ua>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/01/03 14:03:44 by astadnik          #+#    #+#             */
-/*   Updated: 2018/01/05 17:02:34 by astadnik         ###   ########.fr       */
+/*   Updated: 2018/01/06 11:03:40 by astadnik         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ int	ft_printf(const char *format, ...)
 	ret = NULL;
 	rez = printf_make_str(&ret, format, arg, arg_beg);
 	if (~rez)
-		write(1, ret, rez);
+		write(1, ret, (size_t)rez);
 	free(ret);
 	va_end(arg);
 	va_end(arg_beg);
